pyproject.toml
[tool.poetry]
name = "ice_breaker"
version = "0.1.0"
description = ""
authors = ["10138103 <jk.choi@kt.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.32.3"
langchain = "^0.2.11"
black = "^24.4.2"
openai = "^1.37.0"
google-search-results = "^2.4.2"
tweepy = "^4.14.0"
Flask = "^3.0.3"
langchain-openai = "^0.1.17"
langchain-community = "^0.2.10"
langchain-core = "^0.2.23"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
pylint = "^3.2.6"
langchainhub = "^0.1.20"
langchain-ollama = "^0.1.0"
loguru = "^0.7.2"
langsmith = "^0.1.93"
isort = "^5.13.2"
langchain-pinecone = "^0.1.3"
pypdf = "^4.3.1"
faiss-cpu = "^1.8.0"
nltk = "^3.8.1"
uvicorn = "^0.30.3"
fastapi = "^0.111.1"
beautifulsoup4 = "^4.12.3"
tiktoken = "^0.7.0"
unstructured = "^0.15.0"
Jinja2 = "^3.1.4"
streamlit = "^1.37.0"
streamlit-chat = "^0.1.1"
tqdm = "^4.66.4"

[tool.poetry.dev-dependencies]
mypy = "^1.11.0"
debugpy = "^1.8.2"
black = "^24.4.2"
typing-extensions = "^4.12.2"

[tool.mypy]
# --strict
packages = ["ice_breaker", "rag", "react"]
disallow_any_generics = true
disallow_subclassing_any = false
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_return_any = true
implicit_reexport = false
strict_equality = true
ignore_missing_imports = true
# --strict end

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py310']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
  )/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
